feat(agents): convert GithubWizard to Astro-shell StepCards + per-step islands
First component on the Astro-shell house pattern (#12187). The 1500-line single
React wizard island becomes: static StepCard chrome rendered as server HTML +
per-step client:only islands that read guild/tokens from nanostores and drive
their card status imperatively via useStepCardStatus (one data-status write, CSS
palette — no vDOM for the card frames).

- @kbve/astro: wizard/shared.tsx (styles/consts/helpers/CenterMsg), Step1Webhook
  / Step2WebhookConfig / Step3Pat / Step4SmokeBackfill (extracted, props →
  nanostore derivations, StepCard wrapper → useStepCardStatus anchor), and
  WizardGate (auth gating + guild picker + header; reveals the steps container by
  toggling [data-wizard-steps] hidden via closest('[data-wizard-root]')).
- app AstroAgentGithubWizard.astro composes <StepCard> shells around the islands;
  5 thin app wrappers carry the agents-boot side-effect.
- delete the superseded 1500-line ReactAgentGithubWizard monolith + its export
  and the app wrapper (nothing else referenced them).

Build: nx astro-kbve:build green (7742 pages); the github wizard page prerenders
the 4 StepCard frames as static HTML (data-stepcard-root × 4). Interactive
hydration (status pill patching, gate reveal, ClientRouter swap) wants a live
eyeball.
