--- apps/kbve/astro-kbve/src/components/dashboard/agents/ReactAgentGithubWizard.tsx
@@ -1,4 +1,0 @@
-import '@/lib/agents-boot';
-import { ReactAgentGithubWizard } from '@kbve/astro';
-
-export default ReactAgentGithubWizard;
