+++ apps/kbve/astro-kbve/src/components/dashboard/agents/wizard/Step1Webhook.tsx
@@ -1,0 +1,4 @@
+import '@/lib/agents-boot';
+import { Step1Webhook } from '@kbve/astro';
+
+export default Step1Webhook;
