+++ apps/kbve/astro-kbve/src/components/dashboard/agents/wizard/Step2WebhookConfig.tsx
@@ -1,0 +1,4 @@
+import '@/lib/agents-boot';
+import { Step2WebhookConfig } from '@kbve/astro';
+
+export default Step2WebhookConfig;
