+++ apps/kbve/astro-kbve/src/components/dashboard/agents/wizard/Step3Pat.tsx
@@ -1,0 +1,4 @@
+import '@/lib/agents-boot';
+import { Step3Pat } from '@kbve/astro';
+
+export default Step3Pat;
