+++ apps/kbve/astro-kbve/src/components/dashboard/agents/wizard/Step4SmokeBackfill.tsx
@@ -1,0 +1,4 @@
+import '@/lib/agents-boot';
+import { Step4SmokeBackfill } from '@kbve/astro';
+
+export default Step4SmokeBackfill;
