+++ apps/kbve/astro-kbve/src/components/dashboard/agents/wizard/WizardGate.tsx
@@ -1,0 +1,4 @@
+import '@/lib/agents-boot';
+import { WizardGate } from '@kbve/astro';
+
+export default WizardGate;
