+++ packages/npm/astro/src/agents/index.ts
@@ -11,7 +11,11 @@ export { default as ReactAgentGuildPicker } from './ReactAgentGuildPicker';
 export { default as ReactAgentBotConfig } from './ReactAgentBotConfig';
 export { default as ReactAgentRepoAllowlist } from './ReactAgentRepoAllowlist';
 export { default as ReactAgentEventQueue } from './ReactAgentEventQueue';
-export { default as ReactAgentGithubWizard } from './ReactAgentGithubWizard';
+export { default as WizardGate } from './wizard/WizardGate';
+export { default as Step1Webhook } from './wizard/Step1Webhook';
+export { default as Step2WebhookConfig } from './wizard/Step2WebhookConfig';
+export { default as Step3Pat } from './wizard/Step3Pat';
+export { default as Step4SmokeBackfill } from './wizard/Step4SmokeBackfill';
 export { default as ReactAgentSetupStatus } from './ReactAgentSetupStatus';
 export { default as ReactAgentDiscordsh } from './ReactAgentDiscordsh';
 export { default as ReactAgentBotInstall } from './ReactAgentBotInstall';
